Update the template to reflect the changes in the SamplePlugin
Use the new BaseStreamDeckAction class
Update the README with information on creating a new action
